fix dropzone accessibility issue
This just makes some CSS tweaks to allow the user to tab into the hidden
input which is used for the resume upload in the dropzone.

for the input on the dropzone I found a CSS only way to do it.
Basically, input elements should normally be tab-able, but if they
have display: none set on them they get taken out of the tab order.
So basically I removed display: none and simulated it by using
position: fixed and opacity: 0 - this makes the input tab-able.

part of #901
pr #912